requirements.txt
# API
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
aiofiles

# Agent
mistralai
groq
discord.py

# Utilities
pydantic
